starting fresh with network code
