+++ arena.html
@@ -7,6 +7,7 @@
     <script src="//cdn.jsdelivr.net/npm/phaser@3.20.1/dist/phaser.min.js"></script>
     <script src="./lib/phaser-matter-collision-plugin.min.js"></script>
     <script src="./lib/rexvirtualjoystickplugin.min.js"></script>
+    <script src="./lib/phaser-input.min.js"></script>
     
     <!--Initializing Components-->
     <script type="module" src="./src/components/TileMap.js"></script>
