+++ src/entities/ArenaPlayer.js
@@ -91,8 +91,6 @@ export class ArenaPlayer extends Phaser.Physics.Matter.Sprite
 
         this.displayHealth = scene.add.text(30, 12, this.status.health, {color:'#DC143C'});
         this.displayHealth.setScrollFactor(0, 0);
-
-        this.isDead = false;
     }
 
     update()
@@ -254,8 +252,6 @@ export class ArenaPlayer extends Phaser.Physics.Matter.Sprite
 
     //Initializing death sequence
     death() {
-        this.isDead = true;
-
         // Event listeners
         if (this.scene.matter.world) {
             this.scene.matter.world.off("beforeupdate", this.resetTouching, this);
